--- src/core/lib/event_engine/channel_args.h
@@ -1,26 +1,0 @@
-/*
- *
- * Copyright 2021 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-#ifndef GRPC_CORE_LIB_EVENT_ENGINE_CHANNEL_ARGS_H
-#define GRPC_CORE_LIB_EVENT_ENGINE_CHANNEL_ARGS_H
-
-namespace grpc_io {
-// TODO(hork): define
-class ChannelArgs;
-}  // namespace grpc_io
-
-#endif  // GRPC_CORE_LIB_EVENT_ENGINE_CHANNEL_ARGS_H
